Emit Stream frame length correctly
The property `quic.stream.length` that the tool uses doesn't seem to be emitted
by wireshark for frames that carry no explicit length information (stream frame is
at the end of the packet). In that case the generated qlog will not have a length field.

To fix this, we calculate the length based on the data in the frame (which is correctly
emitted).

Fixes #7